refactor: moved TritonServerServiceAbs and Encryption utils to internal package
