+++ kura/org.eclipse.kura.ai.triton.server/src/main/java/org/eclipse/kura/ai/triton/server/TritonServerServiceNativeImpl.java
@@ -13,28 +13,29 @@
 package org.eclipse.kura.ai.triton.server;
 
 import org.eclipse.kura.ai.triton.server.internal.TritonServerInstanceManager;
+import org.eclipse.kura.ai.triton.server.internal.TritonServerServiceAbs;
 import org.eclipse.kura.executor.CommandExecutorService;
 
 public class TritonServerServiceNativeImpl extends TritonServerServiceAbs {
 
     @Override
-    TritonServerInstanceManager createInstanceManager(TritonServerServiceOptions options,
+    protected TritonServerInstanceManager createInstanceManager(TritonServerServiceOptions options,
             CommandExecutorService executorService, String decryptionFolderPath) {
         return new TritonServerNativeManager(options, executorService, decryptionFolderPath);
     }
 
     @Override
-    boolean isConfigurationValid() {
+    protected boolean isConfigurationValid() {
         return !isNullOrEmpty(this.options.getBackendsPath()) && !isNullOrEmpty(this.options.getModelRepositoryPath());
     }
 
     @Override
-    boolean isModelEncryptionEnabled() {
+    protected boolean isModelEncryptionEnabled() {
         return this.options.isModelEncryptionPasswordSet();
     }
 
     @Override
-    String getServerAddress() {
+    protected String getServerAddress() {
         return "localhost";
     }
 }
